add desktop entries: france24/japanews24 live
These live streams both use mpv and youtube-dl. Format is hardcoded as
option 95 which seems to be what it's always called, hopefully it
shouldn't break anytime soon (unless the URLs change, but I don't think
24 hour live stream URLs like these ever change).

An image of an mpv logo is also added to the repo and played before both
streams. This is to have the mpv window show up immediately regardless
how long it takes to load the stream.